+++ foundation/eclipselink.core.test/src/it/java/org/eclipse/persistence/testing/tests/distributedservers/rcm/InvalidateObjectWithMissingReferenceTest.java
@@ -44,8 +44,8 @@ public class InvalidateObjectWithMissingReferenceTest extends ConfigurableCacheS
     public InvalidateObjectWithMissingReferenceTest() {
         super();
         setName("InvalidateObjectWithMissingReferenceTest");
-        cacheSyncConfigValues.put(Employee.class, Integer.valueOf(ClassDescriptor.SEND_OBJECT_CHANGES));
-        cacheSyncConfigValues.put(Address.class, Integer.valueOf(ClassDescriptor.DO_NOT_SEND_CHANGES));
+        cacheSyncConfigValues.put(Employee.class, ClassDescriptor.SEND_OBJECT_CHANGES);
+        cacheSyncConfigValues.put(Address.class, ClassDescriptor.DO_NOT_SEND_CHANGES);
     }
 
     @Override
@@ -59,7 +59,7 @@ public class InvalidateObjectWithMissingReferenceTest extends ConfigurableCacheS
         while (keys.hasMoreElements()) {
             Class keyClass = (Class)keys.nextElement();
             ClassDescriptor descriptor = getSession().getDescriptor(keyClass);
-            int newCacheSyncType = ((Integer)oldCacheSyncConfigValues.get(keyClass)).intValue();
+            int newCacheSyncType = (Integer) oldCacheSyncConfigValues.get(keyClass);
             descriptor.setCacheSynchronizationType(newCacheSyncType);
         }
     }
@@ -76,8 +76,8 @@ public class InvalidateObjectWithMissingReferenceTest extends ConfigurableCacheS
                 int cacheSyncType = descriptor.getCacheSynchronizationType();
                 Object newCacheSyncType = cacheSyncConfigValues.get(keyClass);
                 if (newCacheSyncType != null) {
-                    oldCacheSyncConfigValues.put(keyClass, Integer.valueOf(cacheSyncType));
-                    descriptor.setCacheSynchronizationType(((Integer)newCacheSyncType).intValue());
+                    oldCacheSyncConfigValues.put(keyClass, cacheSyncType);
+                    descriptor.setCacheSynchronizationType((Integer) newCacheSyncType);
                 }
             }
         }
