+++ foundation/org.eclipse.persistence.corba/src/main/java/org/eclipse/persistence/internal/sessions/coordination/corba/sun/_SunCORBAConnectionImplBase.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -38,7 +38,7 @@ public abstract class _SunCORBAConnectionImplBase extends ObjectImpl implements 
     private static Hashtable _methods = new Hashtable();
 
     static {
-        _methods.put("executeCommand", Integer.valueOf(0));
+        _methods.put("executeCommand", 0);
     }
 
     @Override
@@ -49,7 +49,7 @@ public abstract class _SunCORBAConnectionImplBase extends ObjectImpl implements 
             throw new BAD_OPERATION(0, CompletionStatus.COMPLETED_MAYBE);
         }
 
-        switch (__method.intValue()) {
+        switch (__method) {
         case 0:// org/eclipse/persistence/internal/remotecommand/corba/sun/SunCORBAConnection/executeCommand
          {
             byte[] commandData = org.eclipse.persistence.internal.sessions.coordination.corba.sun.CommandDataHelper.read(in);
